Update CalculatePID.ino
This function is used to calculate PID (position, integral,  derivative) and used tested weighting values for each to track and precisely alter movements of servos and motors. This helps to make sure that the two motors responsible for driving are kept instep with each other and do not cause the robot to sway or variate from its path.