Update CheckObjectColor.ino
This code is a combination of both a color checker and stepper motor control. We designed and printed our robot to be made out of black plastic. Through this we are first able to check whether an object is present by checking the brightness index of the chamber. Then, only when an object is detected, we compare the values of each of the r g and b indexes to determine if the object is the correct color. If the object is the correct color, then the stepper motor is activated to drop down a preset amount and drop the object into the launch chamber.  If the object does not match the desired color, the object will continue through the feed chamber and return to the ground of which it came.