Update Definitions.ino
This code includes many of the libraries which we used through the term on various lab exercises. Instead of using the motor control functions available through the MSE2202Lib, we instead opted to create our own motor control system in order to be able to implement PID and PWM signal control to our motors. In addition to this, there are many constant declarations and pin locations defined within this section which are necessary for the complete function of our robot.